requirements.txt
easyAI==2.0.12
numpy==2.1.2

pip~=23.2.1
scikit-fuzzy~=0.5.0
pillow~=11.0.0
pygame~=2.6.1
packaging==24.1
networkx==3.4.2
scipy==1.14.1
pandas~=2.2.3
scikit-learn~=1.5.2
requests~=2.32.3
python-dotenv~=1.0.1